requirements.txt
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.2.0
future==0.18.2
tqdm==4.60.0
gdown==3.12.2
opencv-python==4.4.0.44
glob2==0.7
torch==1.8.1
torchvision==0.9.1
tensorboard==2.5.0